+++ doorstop/core/__init__.py
@@ -3,3 +3,4 @@
 from doorstop.core.item import Item
 from doorstop.core.document import Document
 from doorstop.core.tree import Tree, build, find_document, find_item
+from doorstop.core import importer